put pdf text recognition and related classes in it's own file
put pdf text recognition and related classes in it's own file
tidy up some comments and single line ifs and elses
