+++ win32/msvc2019/pdfimport/pdfimport.vcxproj
@@ -207,6 +207,7 @@
     <ClCompile Include="..\..\..\scribus\plugins\import\pdf\importpdf.cpp" />
     <ClCompile Include="..\..\..\scribus\plugins\import\pdf\importpdfplugin.cpp" />
     <ClCompile Include="..\..\..\scribus\plugins\import\pdf\pdfimportoptions.cpp" />
+    <ClCompile Include="..\..\..\scribus\plugins\import\pdf\PdfTextRecognition.cpp" />
     <ClCompile Include="..\..\..\scribus\plugins\import\pdf\slaoutput.cpp" />
     <ClCompile Include="..\..\..\scribus\plugins\plugins_pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
@@ -217,6 +218,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\..\scribus\plugins\import\pdf\importpdfconfig.h" />
+    <ClInclude Include="..\..\..\scribus\plugins\import\pdf\PdfTextRecognition.h" />
     <ClInclude Include="..\..\..\scribus\plugins\plugins_pch.h" />
     <moc Include="..\..\..\scribus\plugins\import\pdf\importpdf.h" />
     <moc Include="..\..\..\scribus\plugins\import\pdf\importpdfplugin.h" />
